requirements.txt
pydantic>=2.0
pandas
numpy
scikit-learn
textblob
vaderSentiment
google-genai
streamlit
pytest
python-dotenv
plotly
